Windows port: experimental padded_slice: fix narrowing in kernel args
- padded_slice_rm_program_factory.cpp: static_cast<uint32_t> the int
  input_page_size and this_input_row_size_bytes values placed into
  uint32_t kernel-arg vectors (C2398 narrowing).
- padded_slice_tile_program_factory.cpp: static_cast<uint32_t> the
  size_t input_padded_shape.rank() value (C2398 narrowing).
- plusone and slice_write targets already compiled clean.

+++ ttnn/cpp/ttnn/operations/experimental/padded_slice/device/padded_slice_rm_program_factory.cpp
@@ -108,7 +108,7 @@ get_padded_slice_runtime_args_rm_sharded_output(
     uint32_t start_addr = input_tensor.buffer()->address();
     std::vector<uint32_t> common_reader_kernel_args = {
         start_addr + begins_bytes - misalignment,  // read from nearest aligned address
-        input_page_size,
+        static_cast<uint32_t>(input_page_size),
         output_row_size_bytes,
         output_row_size_bytes_offset,
         num_dims,
@@ -182,7 +182,10 @@ get_padded_slice_runtime_args_rm_sharded_output(
             this_input_row_size_bytes);
 
         std::vector<uint32_t> writer_kernel_args = {
-            this_core_num_sticks, output_row_size_elems, this_input_row_size_bytes, output_row_size_bytes};
+            this_core_num_sticks,
+            output_row_size_elems,
+            static_cast<uint32_t>(this_input_row_size_bytes),
+            output_row_size_bytes};
         ret_val[core_index] = {reader_kernel_args, writer_kernel_args};
         core_index++;
     }
